+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/PublishingPlugin.kt
@@ -14,6 +14,7 @@
 
 package com.google.firebase.gradle.plugins
 
+import com.google.common.collect.Sets
 import com.google.firebase.gradle.bomgenerator.BomGeneratorTask
 import com.google.firebase.gradle.plugins.PublishingPlugin.Companion.BUILD_BOM_ZIP_TASK
 import com.google.firebase.gradle.plugins.PublishingPlugin.Companion.BUILD_KOTLINDOC_ZIP_TASK
@@ -30,6 +31,7 @@ import org.gradle.api.Project
 import org.gradle.api.publish.maven.tasks.PublishToMavenRepository
 import org.gradle.api.tasks.Copy
 import org.gradle.api.tasks.bundling.Zip
+import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.register
 
@@ -60,6 +62,9 @@ import org.gradle.kotlin.dsl.register
  * - [PUBLISH_RELEASING_LIBS_TO_LOCAL_TASK][registerPublishReleasingLibrariesToMavenLocalTask]
  * - [SEMVER_CHECK_TASK][registerSemverCheckForReleaseTask]
  * - [PUBLISH_ALL_TO_BUILD_TASK][registerPublishAllToBuildDir]
+ *
+ * Additionally, this plugin registers the [PostReleasePlugin] via [registerPostReleasePlugin] for
+ * each releasing library.
  */
 abstract class PublishingPlugin : Plugin<Project> {
   override fun apply(project: Project) {
@@ -75,7 +80,7 @@ abstract class PublishingPlugin : Plugin<Project> {
       val checkHeadDependencies =
         registerCheckHeadDependenciesTask(project, releasingFirebaseLibraries)
       val validateProjectsToPublish =
-        registerValidateProjectsToPublishTask(project, releasingProjects)
+        registerValidateProjectsToPublishTask(project, releasingFirebaseLibraries)
       val publishReleasingLibrariesToBuildDir =
         registerPublishReleasingLibrariesToBuildDirTask(project, releasingProjects)
       val generateKotlindocsForRelease =
@@ -91,6 +96,7 @@ abstract class PublishingPlugin : Plugin<Project> {
       registerPublishReleasingLibrariesToMavenLocalTask(project, releasingProjects)
       registerSemverCheckForReleaseTask(project, releasingProjects)
       registerPublishAllToBuildDir(project, allFirebaseLibraries)
+      registerPostReleasePlugin(releasingProjects)
 
       val buildMavenZip =
         project.tasks.register<Zip>(BUILD_MAVEN_ZIP_TASK) {
@@ -270,15 +276,24 @@ abstract class PublishingPlugin : Plugin<Project> {
   // TODO(b/280320915): Remove doLast when Gradle + IDEA fix task configuration avoidance bug
   private fun registerValidateProjectsToPublishTask(
     project: Project,
-    releasingProjects: List<Project>
+    releasinglibraries: List<FirebaseLibraryExtension>
   ) =
     project.tasks.register(VALIDATE_PROJECTS_TO_PUBLISH_TASK) {
       doLast {
-        if (releasingProjects.isEmpty()) {
+        if (releasinglibraries.isEmpty()) {
           throw GradleException(
             "No projects to release. " +
               "Ensure you've specified the projectsToPublish parameter, " +
               "or have a valid $RELEASE_CONFIG_FILE file at the root directory."
+          )
+        }
+        val libraryGroupProjects =
+          releasinglibraries.flatMap { it.projectsToRelease }.filterNotNull().toSet()
+        val releasingProjects = releasinglibraries.mapNotNull { it.project }.toSet()
+        if (!libraryGroupProjects.equals(releasingProjects)) {
+          throw GradleException(
+            "Some libraries in library groups are not in the release: " +
+              Sets.difference(libraryGroupProjects, releasingProjects).map { it.displayName }
           )
         }
       }
@@ -373,6 +388,7 @@ abstract class PublishingPlugin : Plugin<Project> {
       currentRelease.convention(project.provideProperty("currentRelease"))
       pastRelease.convention(project.provideProperty("pastRelease"))
       printReleaseConfig.convention(project.provideProperty("printOutput"))
+      commitsToIgnoreFile.convention(project.layout.projectDirectory.file("ignoreCommits.txt"))
 
       releaseConfigFile.convention(project.layout.projectDirectory.file(RELEASE_CONFIG_FILE))
       releaseReportMdFile.convention(project.layout.projectDirectory.file(RELEASE_REPORT_MD_FILE))
@@ -444,6 +460,13 @@ abstract class PublishingPlugin : Plugin<Project> {
       }
     }
 
+  /** Registers the [PostReleasePlugin] to each releaing project. */
+  private fun registerPostReleasePlugin(releasingProjects: List<Project>) {
+    for (releasingProject in releasingProjects) {
+      releasingProject.apply<PostReleasePlugin>()
+    }
+  }
+
   companion object {
     const val RELEASE_CONFIG_FILE = "release.json"
     const val RELEASE_REPORT_MD_FILE = "release_report.md"
@@ -476,7 +499,6 @@ abstract class PublishingPlugin : Plugin<Project> {
  *
  * @property releasingLibraries A list of libraries that should be released
  * @property name The name of the release (such as `m123`)
- *
  * @see computeReleaseMetadata
  */
 data class ReleaseMetadata(