+++ firebase-firestore/src/main/java/com/google/firebase/firestore/FirebaseFirestore.java
@@ -23,6 +23,7 @@ import android.content.Context;
 import androidx.annotation.Keep;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RestrictTo;
 import androidx.annotation.VisibleForTesting;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.TaskCompletionSource;
@@ -104,6 +105,8 @@ public class FirebaseFirestore {
   private volatile FirestoreClient client;
   private final GrpcMetadataProvider metadataProvider;
 
+  @Nullable private PersistentCacheIndexManager persistentCacheIndexManager;
+
   @NonNull
   private static FirebaseApp getDefaultFirebaseApp() {
     FirebaseApp app = FirebaseApp.getInstance();
@@ -114,9 +117,10 @@ public class FirebaseFirestore {
   }
 
   /**
-   * Returns the default {@link FirebaseFirestore} instance associated with the default {@link
-   * FirebaseApp}. Returns the same instance for all invocations. If no instance exists, initializes
-   * a new instance with default settings.
+   * Returns the default {@link FirebaseFirestore} instance for the default {@link FirebaseApp}.
+   *
+   * <p>Returns the same instance for all invocations. If no instance exists, initializes a new
+   * instance.
    *
    * @returns The {@link FirebaseFirestore} instance.
    */
@@ -126,11 +130,12 @@ public class FirebaseFirestore {
   }
 
   /**
-   * Returns the default {@link FirebaseFirestore} instance that is associated with the provided
-   * {@link FirebaseApp}. For a given {@link FirebaseApp}, invocation always returns the same
-   * instance. If no instance exists, initializes a new instance with default settings.
+   * Returns the default {@link FirebaseFirestore} instance for the provided {@link FirebaseApp}.
    *
-   * @param app - The {@link FirebaseApp} instance that the returned {@link FirebaseFirestore}
+   * <p>For a given {@link FirebaseApp}, invocation always returns the same instance. If no instance
+   * exists, initializes a new instance.
+   *
+   * @param app The {@link FirebaseApp} instance that the returned {@link FirebaseFirestore}
    *     instance is associated with.
    * @returns The {@link FirebaseFirestore} instance.
    */
@@ -140,11 +145,12 @@ public class FirebaseFirestore {
   }
 
   /**
-   * Returns the {@link FirebaseFirestore} instance that is associated with the default {@link
-   * FirebaseApp}. Returns the same instance for all invocations given the same database parameter.
-   * If no instance exists, initializes a new instance with default settings.
+   * Returns the {@link FirebaseFirestore} instance for the default {@link FirebaseApp}.
    *
-   * @param database - The name of database.
+   * <p>Returns the same instance for all invocations given the same database parameter. If no
+   * instance exists, initializes a new instance.
+   *
+   * @param database The database ID.
    * @returns The {@link FirebaseFirestore} instance.
    */
   @NonNull
@@ -153,14 +159,14 @@ public class FirebaseFirestore {
   }
 
   /**
-   * Returns the {@link FirebaseFirestore} instance that is associated with the provided {@link
-   * FirebaseApp}. Returns the same instance for all invocations given the same {@link FirebaseApp}
-   * and database parameter. If no instance exists, initializes a new instance with default
-   * settings.
+   * Returns the {@link FirebaseFirestore} instance for the provided {@link FirebaseApp}.
    *
-   * @param app - The {@link FirebaseApp} instance that the returned {@link FirebaseFirestore}
+   * <p>Returns the same instance for all invocations given the same {@link FirebaseApp} and
+   * database parameter. If no instance exists, initializes a new instance.
+   *
+   * @param app The {@link FirebaseApp} instance that the returned {@link FirebaseFirestore}
    *     instance is associated with.
-   * @param database - The name of database.
+   * @param database The database ID.
    * @returns The {@link FirebaseFirestore} instance.
    */
   @NonNull
@@ -398,6 +404,26 @@ public class FirebaseFirestore {
     }
 
     return client.configureFieldIndexes(parsedIndexes);
+  }
+
+  /**
+   * Returns the PersistentCache Index Manager used by this {@code FirebaseFirestore} object.
+   *
+   * @return The {@code PersistentCacheIndexManager} instance or null if local persistent storage is
+   *     not in use.
+   */
+  // TODO(csi): Remove the `hide` and scope annotations.
+  /** @hide */
+  @RestrictTo(RestrictTo.Scope.LIBRARY)
+  @Nullable
+  public synchronized PersistentCacheIndexManager getPersistentCacheIndexManager() {
+    ensureClientConfigured();
+    if (persistentCacheIndexManager == null
+        && (settings.isPersistenceEnabled()
+            || settings.getCacheSettings() instanceof PersistentCacheSettings)) {
+      persistentCacheIndexManager = new PersistentCacheIndexManager(client);
+    }
+    return persistentCacheIndexManager;
   }
 
   /**