+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/BloomFilter.java
@@ -16,7 +16,6 @@ package com.google.firebase.firestore.remote;
 
 import android.util.Base64;
 import androidx.annotation.NonNull;
-import androidx.annotation.VisibleForTesting;
 import com.google.protobuf.ByteString;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
@@ -93,7 +92,6 @@ public final class BloomFilter {
     }
   }
 
-  @VisibleForTesting
   int getBitCount() {
     return this.bitCount;
   }