+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/WatchChangeAggregator.java
@@ -85,7 +85,7 @@ public class WatchChangeAggregator {
   private static final String LOG_TAG = "WatchChangeAggregator";
 
   /** The bloom filter application status while handling existence filter mismatch. */
-  private enum BloomFilterApplicationStatus {
+  enum BloomFilterApplicationStatus {
     SUCCESS,
     SKIPPED,
     FALSE_POSITIVE
@@ -216,7 +216,11 @@ public class WatchChangeAggregator {
         if (currentSize != expectedCount) {
 
           // Apply bloom filter to identify and mark removed documents.
-          BloomFilterApplicationStatus status = this.applyBloomFilter(watchChange, currentSize);
+          BloomFilter bloomFilter = this.parseBloomFilter(watchChange);
+          BloomFilterApplicationStatus status =
+              bloomFilter != null
+                  ? this.applyBloomFilter(bloomFilter, watchChange, currentSize)
+                  : BloomFilterApplicationStatus.SKIPPED;
 
           if (status != BloomFilterApplicationStatus.SUCCESS) {
             // If bloom filter application fails, we reset the mapping and
@@ -234,28 +238,27 @@ public class WatchChangeAggregator {
           TestingHooks.getInstance()
               .notifyOnExistenceFilterMismatch(
                   TestingHooks.ExistenceFilterMismatchInfo.from(
-                      status == BloomFilterApplicationStatus.SUCCESS,
                       currentSize,
-                      watchChange.getExistenceFilter()));
+                      watchChange.getExistenceFilter(),
+                      targetMetadataProvider.getDatabaseId(),
+                      bloomFilter,
+                      status));
         }
       }
     }
   }
 
-  /** Apply bloom filter to remove the deleted documents, and return the application status. */
-  private BloomFilterApplicationStatus applyBloomFilter(
-      ExistenceFilterWatchChange watchChange, int currentCount) {
-    int expectedCount = watchChange.getExistenceFilter().getCount();
+  /** Parse the bloom filter from the "unchanged_names" field of an existence filter. */
+  @Nullable
+  private BloomFilter parseBloomFilter(ExistenceFilterWatchChange watchChange) {
     com.google.firestore.v1.BloomFilter unchangedNames =
         watchChange.getExistenceFilter().getUnchangedNames();
-
     if (unchangedNames == null || !unchangedNames.hasBits()) {
-      return BloomFilterApplicationStatus.SKIPPED;
+      return null;
     }
 
     ByteString bitmap = unchangedNames.getBits().getBitmap();
     BloomFilter bloomFilter;
-
     try {
       bloomFilter =
           BloomFilter.create(
@@ -266,20 +269,26 @@ public class WatchChangeAggregator {
           "Applying bloom filter failed: ("
               + e.getMessage()
               + "); ignoring the bloom filter and falling back to full re-query.");
-      return BloomFilterApplicationStatus.SKIPPED;
+      return null;
     }
 
     if (bloomFilter.getBitCount() == 0) {
-      return BloomFilterApplicationStatus.SKIPPED;
+      return null;
     }
+
+    return bloomFilter;
+  }
+
+  /** Apply bloom filter to remove the deleted documents, and return the application status. */
+  private BloomFilterApplicationStatus applyBloomFilter(
+      BloomFilter bloomFilter, ExistenceFilterWatchChange watchChange, int currentCount) {
+    int expectedCount = watchChange.getExistenceFilter().getCount();
 
     int removedDocumentCount = this.filterRemovedDocuments(bloomFilter, watchChange.getTargetId());
 
-    if (expectedCount != (currentCount - removedDocumentCount)) {
-      return BloomFilterApplicationStatus.FALSE_POSITIVE;
-    }
-
-    return BloomFilterApplicationStatus.SUCCESS;
+    return (expectedCount == (currentCount - removedDocumentCount))
+        ? BloomFilterApplicationStatus.SUCCESS
+        : BloomFilterApplicationStatus.FALSE_POSITIVE;
   }
 
   /**